test: verify current wallet retrieval flow
